Task Force Spartan Resolve
The first US Army conventional unit to deploy to counter the Russian invasion is the 173rd Airborne Brigade Combat Team, the famous “Sky Soldiers” based in Vicenza, Italy. The deployment of the 173rd Airborne (as Task Force Spartan Resolve) is conducted with the high hope that they will be the keystone of a physical “red line” that the Russian High Command will not dare to cross in central Ukraine. This campaign centers around the actions of “Attack” Company, 1-503rd Airborne, as they are rushed to the critical Cherkasy Area of Operations (AO) to reinforce the Ukrainian 30th Mechanized Brigade.

Notes:
This is the third (and final!) version of this campaign (v3.0). Corrections were made to the scoring in some of the scenarios, as well as updated AI plans.