Add files via upload
Author: IanL

This is the map used for the scenario that ships with CMFB “Action at Lanzerath Ridge”.  I made the map using google earth (or the terrain) and the pictures and maps in the book The Longest Winter  https://www.amazon.ca/The-Longest-Winter-Decorated-Platoon/dp/0306814404 (for period pictures).  This map is much larger than what was used in the scenario.  It includes the ridge where the fighting took place, the village and a fair amount of the area south of the village too.